fixing initial scalings in marginal_screening
